requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.112.2
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
idna==3.8
Mako==1.3.5
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
strawberry-graphql==0.237.3
typing_extensions==4.12.2
uvicorn==0.30.6
